test(Product): 상품 생성 RestAssured 기반 테스트로 변경
